+++ src/index.html
@@ -48,11 +48,11 @@
       <option value="Fire">Fuego</option>
       <option value="Ice">Hielo</option>
       <option value="Normal">Normal</option>
-      <option value="Fighting">Pelea</option>
+      <option value="Fighting">Lucha</option>
       <option value="Grass">Planta</option>
       <option value="Psychic">Psíquicos</option>
       <option value="Rock">Roca</option>
-      <option alue="Ground">Tierra</option>
+      <option value="Ground">Tierra</option>
       <option value="Poison">Veneno</option>
       <option value="Flying">Volador</option>
     </select>
